Add ion contributions to SDEC plotter (#1539)
* Add catch for missing elements

The elements coloured in the plot are determined via the absorption and emission. It's possible however, that these elements are not the same, e.g. if you specify a limited wavelength range. This could give an error if you try to plot an element not in the emission dataframe, for example. This adds a try statement to catch this case

* Fixed typo

* Black formatting

* Fixed typo

* Add species column to packet df

Added a new column to the packet dataframe that gives the species id. This combines the atomic and ion numbers into a unique id that will be used to group species later.

* Added species list parser

Added parser to convert request species list into ids to be used by plotter

* Implement parser call

Call species list parser function

* Fixed parser

Rearranged parser so it didn't crash if species_list was none

* Implement MPL ion contributions

Added in the ion contribution stuff for the mpl plotting.
Also added a new function to get the colorbar labels. This is necessary to determine how many unique colours there should be

* Changed order of functions

* Implement plotly ion contributions

Added colouring for ion contributions in plotly figures

* Remove print statements

* Fixed df bug and added comments

Fixed bug in the absorption and emission df calculations. Now they elements should be handled correctly so there won't be errors if a species appears in emission and not absorption, and vice versa

Added comments throughout

* Black formatting

* Added colour function

Added a new function to work out what colours will be plotted. This was originally repeated multiple times in the code, but now it's a single function that's called multiple times

Also fixed a bug in the absorption luminosities dataframe

* Address review comments

Change name of some variables and fixed some typos.

* Change print to logger

* Black formatting

* Updated variable names and string formatting

* Black formatting